pyproject.toml
[tool.poetry]
name = "weather-forecast-scraper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "weather_forecast_scraper"}]

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.1.1"
pytest = "^8.3.3"
requests = "^2.32.3"
environs = "^11.0.0"
scrapy = "^2.11.2"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.2"
matplotlib = "^3.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
